requirements.txt
absl-py==2.1.0
alembic==1.13.2
aniso8601==9.0.1
asttokens==2.4.1
astunparse==1.6.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.32.3
debugpy==1.8.6
decorator==5.1.1
Deprecated==1.2.14
docker==7.1.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.2.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.66.1
h5py==3.11.0
idna==3.10
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
logging==0.4.9.6
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mlflow==2.16.2
mlflow-skinny==2.16.2
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.8.1
rsa==4.9
schedule==1.2.2
scheduler==0.8.7
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tornado==6.4.1
traitlets==5.14.3
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
